refactor: move plugin-specific config into extensions
