+++ extensions/minimax/index.ts
@@ -8,13 +8,10 @@ import {
 } from "openclaw/plugin-sdk/minimax-portal-auth";
 import { ensureAuthProfileStore, listProfilesForProvider } from "../../src/agents/auth-profiles.js";
 import { MINIMAX_OAUTH_MARKER } from "../../src/agents/model-auth-markers.js";
-import {
-  applyMinimaxApiConfig,
-  applyMinimaxApiConfigCn,
-} from "../../src/commands/onboard-auth.config-minimax.js";
 import { fetchMinimaxUsage } from "../../src/infra/provider-usage.fetch.js";
 import { createProviderApiKeyAuthMethod } from "../../src/plugins/provider-api-key-auth.js";
 import { loginMiniMaxPortalOAuth, type MiniMaxRegion } from "./oauth.js";
+import { applyMinimaxApiConfig, applyMinimaxApiConfigCn } from "./onboard.js";
 import { buildMinimaxPortalProvider, buildMinimaxProvider } from "./provider-catalog.js";
 
 const API_PROVIDER_ID = "minimax";
