+++ extensions/opencode-go/index.ts
@@ -1,7 +1,7 @@
 import { emptyPluginConfigSchema, type OpenClawPluginApi } from "openclaw/plugin-sdk/core";
-import { applyOpencodeGoConfig } from "../../src/commands/onboard-auth.config-opencode-go.js";
 import { OPENCODE_GO_DEFAULT_MODEL_REF } from "../../src/commands/opencode-go-model-default.js";
 import { createProviderApiKeyAuthMethod } from "../../src/plugins/provider-api-key-auth.js";
+import { applyOpencodeGoConfig } from "./onboard.js";
 
 const PROVIDER_ID = "opencode-go";
